(1) Add support for OpenAI GPT4.5; (2) Solved a subtle bug, see description for more
``` bug report
Background:
We detect and ignore the trailing newline characters (as a workaround to achieve scrolling beyond the last line)
Problem:
However, the current implementation re-detects the number of newline characters if there's a break in the animation process. Meaning, for slower language models such as GPT4.5, the newline characters generated by the language model will be ignored. This is not acceptable, as it breaks the structures of many responses. The solution is to detect only ONCE for each generation process. The rationale behind this is that, the text editor is set to read-only during generation, so the number of trailing newline characters will not change.
```